+++ centurion-avro/src/main/kotlin/com/sksamuel/centurion/avro/io/Serde.kt
@@ -1,0 +1,42 @@
+package com.sksamuel.centurion.avro.io
+
+import com.sksamuel.centurion.avro.decoders.SpecificRecordDecoder
+import com.sksamuel.centurion.avro.encoders.RecordDecoder
+import com.sksamuel.centurion.avro.encoders.RecordEncoder
+import com.sksamuel.centurion.avro.encoders.SpecificRecordEncoder
+import com.sksamuel.centurion.avro.generation.ReflectionSchemaBuilder
+import org.apache.avro.Schema
+import org.apache.avro.io.EncoderFactory
+import java.io.ByteArrayInputStream
+import kotlin.reflect.KClass
+
+/**
+ * A [Serde] provides an easy way to convert between data classes and avro encoded bytes.
+ */
+class Serde<T : Any>(schema: Schema, kclass: KClass<T>) {
+
+   init {
+      require(kclass.isData)
+   }
+
+   companion object {
+
+      operator fun <T : Any> invoke(kclass: KClass<T>): Serde<T> {
+         val schema = ReflectionSchemaBuilder(true).schema(kclass)
+         return Serde(schema, kclass)
+      }
+
+      inline operator fun <reified T : Any> invoke(): Serde<T> {
+         val schema = ReflectionSchemaBuilder(true).schema(T::class)
+         return Serde(schema, T::class)
+      }
+   }
+
+   private val encoder = RecordEncoder(schema, SpecificRecordEncoder(kclass, schema))
+   private val decoder = RecordDecoder(SpecificRecordDecoder(kclass, schema))
+   private val writerFactory = AvroBinaryWriterFactory(schema, EncoderFactory())
+   private val readerFactory = AvroBinaryReaderFactory(schema)
+
+   fun serialize(obj: T): ByteArray = writerFactory.writer().write(encoder.encode(obj)).bytes()
+   fun deserialize(bytes: ByteArray): T = decoder.decode((readerFactory.reader(ByteArrayInputStream(bytes)).read()))
+}